+++ samples/stopwatch-compose/src/main/java/com/instacart/formula/compose/stopwatch/StopwatchFeatureFactory.kt
@@ -23,8 +23,8 @@ import com.instacart.formula.android.compose.ComposeViewFactory
 import com.instacart.formula.invoke
 import com.instacart.formula.rxjava3.toObservable
 
-class StopwatchFeatureFactory : FeatureFactory<Any, StopwatchKey> {
-    override fun initialize(dependencies: Any, key: StopwatchKey): Feature {
+class StopwatchFeatureFactory : FeatureFactory<Any, StopwatchKey>() {
+    override fun Params.initialize(): Feature {
         return Feature(
             state = StopwatchFormula().toObservable(),
             viewFactory = StopwatchViewFactory()