+++ samples/todoapp/src/main/java/com/examples/todoapp/tasks/TaskListFeatureFactory.kt
@@ -9,13 +9,13 @@ import com.instacart.formula.android.LayoutViewFactory
 import com.instacart.formula.android.ViewInstance
 import com.instacart.formula.rxjava3.toObservable
 
-class TaskListFeatureFactory : FeatureFactory<TaskListFeatureFactory.Dependencies, TaskListKey> {
+class TaskListFeatureFactory : FeatureFactory<TaskListFeatureFactory.Dependencies, TaskListKey>() {
     interface Dependencies {
         fun taskRepo(): TaskRepo
         fun taskListInput(): TaskListFormula.Input
     }
 
-    override fun initialize(dependencies: Dependencies, key: TaskListKey): Feature {
+    override fun Params.initialize(): Feature {
         // Note: we could create our own internal dagger component here using the dependencies.
         val formula = TaskListFormula(dependencies.taskRepo())
         return Feature(